+++ packages/api/src/modules/providers/osiris/@types/IOsirisEvaluationsInformations.ts
@@ -4,5 +4,6 @@ export default interface IOsirisEvaluationsInformations {
     osirisActionId: string,
     siret: Siret,
     evaluation_resultat: string,
+    extractYear: number,
     cout_total_realise?: number
 }