+++ packages/api/src/modules/providers/osiris/@types/IOsirisRequestInformations.ts
@@ -38,4 +38,6 @@ export default interface IOsirisRequestInformations extends IProviderInformation
     versementRealise: number,
     versementCompensationN1: number,
     versementCompensationN: number,
+
+    extractYear: number
 }