+++ packages/api/tests/modules/providers/osiris/osiris.parser.test.ts
@@ -10,7 +10,7 @@ describe("OsirisParser", () => {
     describe('parseRequests', () => {
         it('should return osiris requests', () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/SuiviDossiers_test.xls"));
-            const requests = OsirisParser.parseRequests(buffer);
+            const requests = OsirisParser.parseRequests(buffer, 2022);
 
             expect(requests).toHaveLength(1);
             expect(requests[0]).toBeInstanceOf(OsirisRequestEntity);
@@ -18,7 +18,7 @@ describe("OsirisParser", () => {
 
         it('should have good properties', () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/SuiviDossiers_test.xls"));
-            const requests = OsirisParser.parseRequests(buffer);
+            const requests = OsirisParser.parseRequests(buffer, 2022);
             expect(requests[0]).toMatchObject({
                 legalInformations: {
                     siret: "0",
@@ -37,7 +37,7 @@ describe("OsirisParser", () => {
     describe('parseActions', () => {
         it('should return osiris actions', () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/SuiviActions_test.xls"));
-            const actions = OsirisParser.parseActions(buffer);
+            const actions = OsirisParser.parseActions(buffer, 2022);
 
             expect(actions).toHaveLength(1);
             expect(actions[0]).toBeInstanceOf(OsirisActionEntity);
@@ -45,7 +45,7 @@ describe("OsirisParser", () => {
         
         it('should have good properties', () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/SuiviActions_test.xls"));
-            const actions = OsirisParser.parseActions(buffer);
+            const actions = OsirisParser.parseActions(buffer, 2022);
             expect(actions[0]).toMatchObject({
                 "indexedInformations": {
                     "osirisActionId": "DD00-21-0000-1",
@@ -58,14 +58,14 @@ describe("OsirisParser", () => {
     describe("parseEvaluations()", () => {
         it("should return osiris evaluation", () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/SuiviActionsEvaluation_test.xlsx"));
-            const actual = OsirisParser.parseEvaluations(buffer);
+            const actual = OsirisParser.parseEvaluations(buffer, 2022);
             expect(actual).toHaveLength(1);
             expect(actual[0]).toBeInstanceOf(OsirisEvaluationEntity);
         })
         
         it("should have properties", () => {
             const buffer = fs.readFileSync(path.resolve(__dirname, "./__fixtures__/test.xlsx"));
-            const actual = (OsirisParser.parseEvaluations(buffer)[0]).indexedInformations;
+            const actual = (OsirisParser.parseEvaluations(buffer, 2022)[0]).indexedInformations;
             const expected =  {
                 "osirisActionId": "DD00-21-0000-1",
                 "siret": "0",