+++ packages/front/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.5](https://github.com/betagouv/datasubvention/compare/v0.10.3...v0.10.5) (2022-06-21)
+
+
+### Bug Fixes
+
+* **front:** fix error errorcode is not defined ([#302](https://github.com/betagouv/datasubvention/issues/302)) ([776ff2d](https://github.com/betagouv/datasubvention/commit/776ff2d06681aebd582a16fbfbfceaf8078f0c2a))
+
+
+
+
+
 ## [0.10.4](https://github.com/betagouv/datasubvention/compare/v0.10.3...v0.10.4) (2022-06-14)